Fix/listagem notas

+++ www/conexao_db/notas_crud.php
@@ -2,7 +2,6 @@
 session_start();
 require_once __DIR__ .'/conexao.php';
 
-
 $action = $_POST['action'] ?? null;
 
 try {
@@ -11,7 +10,7 @@ try {
         $descricao = $_POST['descricao'] ?? '';
         $pasta = $_POST['pasta'] ?? '';
         $tipo = $_POST['tipo'] ?? 'Checklist';
-        $id_usuario = $_POST['id_usuario'] ?? 1; 
+        $id_usuario = $_SESSION['id']; 
     
         if ($id_usuario) {
             // Inserir a nota com o ID do usuário
@@ -29,15 +28,20 @@ try {
             echo json_encode(['success' => false, 'message' => 'ID do usuário não fornecido.']);
         }
     } elseif ($action === 'read') {
-        // Ler todas as notas
-        $stmt = $pdo->query("SELECT * FROM notas");
-        $notas = $stmt->fetchAll(PDO::FETCH_ASSOC);
-
+        // Ler todas as notas do usuário
+        $id_usuario = $_SESSION['id'];
+        if($id_usuario){
+            $stmt = $pdo->prepare("SELECT * FROM notas WHERE id_usuario = :id_usuario");
+            $stmt->execute([':id_usuario' => $id_usuario]);
+            $notas = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        } else {
+            echo json_encode(['success' => false, 'message' => 'ID do usuário não encontrado']);
+        }
         echo json_encode($notas);
     } elseif ($action === 'update') {
         // Atualizar uma nota
         $id = $_POST['id'] ?? null; // ID da nota a ser atualizada
-        $id_usuario = $_POST['id_usuario'] ?? 1;
+        $id_usuario = $_SESSION['id'];
         $titulo = $_POST['titulo'] ?? '';
         $descricao = $_POST['descricao'] ?? '';
         $pasta = $_POST['pasta'] ?? '';
@@ -59,10 +63,11 @@ try {
     } elseif ($action === 'delete') {
         // Deletar uma nota
         $id = $_POST['id'];
+        $id_usuario = $_SESSION['id'];
 
-        if ($id) {
+        if ($id && $id_usuario) {
             $stmt = $pdo->prepare("DELETE FROM notas WHERE id = :id AND id_usuario = :id_usuario");
-            $stmt->execute([':id' => $id, ':id_usuario' => 1]);
+            $stmt->execute([':id' => $id, ':id_usuario' => $id_usuario]);
 
             echo json_encode(['success' => true, 'message' => 'Nota deletada com sucesso!']);
         } else {
